# backend/schema/openapi.json
{
    "openapi": "3.0.3",
    "info": {
        "title": "Pinaks API",
        "version": "0.1.0"
    },
    "paths": {
...
        "/api/v1/document-previews/{preview_id}/": {
            "get": {
                "operationId": "document_preview_retrieve",
                "parameters": [
                    {
                        "in": "path",
                        "name": "preview_id",
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        },
                        "required": true
                    }
                ],
                "tags": [
                    "document-previews"
                ],
                "security": [
                    {
                        "cookieAuth": []
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "string",
                                    "format": "binary"
                                }
                            }
                        },
                        "description": ""
                    }
                }
            }
        },
        "/api/v1/document-templates/": {
            "get": {
                "operationId": "document_template_list",
                "parameters": [
                    {
                        "name": "ordering",
                        "required": false,
                        "in": "query",
                        "description": "Which field to use when ordering the results.",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "search",
                        "required": false,
                        "in": "query",
                        "description": "A search term.",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "tags": [
                    "document-templates"
                ],
                "security": [
                    {
                        "cookieAuth": []
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Template"
                                    }
                                }
                            }
                        },
                        "description": ""
                    }
                }
            },
            "post": {
                "operationId": "document_template_create",
                "tags": [
                    "document-templates"
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/TemplateWriteRequest"
                            }
                        },
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "$ref": "#/components/schemas/TemplateWriteRequest"
                            }
                        },
                        "multipart/form-data": {
                            "schema": {
                                "$ref": "#/components/schemas/TemplateWriteRequest"
                            }
                        }
                    },
                    "required": true
                },
                "security": [
                    {
                        "cookieAuth": []
                    }
                ],
                "responses": {
                    "201": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Template"
                                }
                            }
                        },
                        "description": ""
                    },
                    "400": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorEnvelope"
                                }
                            }
                        },
                        "description": ""
                    }
                }
            }
        },
        "/api/v1/document-templates/{template_id}/": {
            "get": {
                "operationId": "document_template_retrieve",
                "parameters": [
                    {
                        "in": "path",
                        "name": "template_id",
                        "schema": {
                            "type": "integer"
                        },
                        "required": true
                    }
                ],
                "tags": [
                    "document-templates"
                ],
                "security": [
                    {
                        "cookieAuth": []
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Template"
                                }
                            }
                        },
                        "description": ""
                    }
                }
            }
        },
        "/api/v1/document-templates/{template_id}/assets/": {
            "post": {
                "operationId": "document_template_asset_upload",
                "parameters": [
                    {
                        "in": "path",
                        "name": "template_id",
                        "schema": {
                            "type": "integer"
                        },
                        "required": true
                    }
                ],
                "tags": [
                    "document-templates"
                ],
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "$ref": "#/components/schemas/AssetUploadRequest"
                            }
                        },
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "$ref": "#/components/schemas/AssetUploadRequest"
                            }
                        }
                    },
                    "required": true
                },
                "security": [
                    {
                        "cookieAuth": []
                    }
                ],
                "responses": {
                    "201": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/DocumentAsset"
                                }
                            }
                        },
                        "description": ""
                    },
                    "400": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorEnvelope"
                                }
                            }
                        },
                        "description": ""
                    }
                }
            }
        },
        "/api/v1/document-templates/{template_id}/preview/": {
            "post": {
                "operationId": "document_template_preview",
                "parameters": [
                    {
                        "in": "path",
                        "name": "template_id",
                        "schema": {
                            "type": "integer"
                        },
                        "required": true
                    }
                ],
                "tags": [
                    "document-templates"
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PreviewRequestRequest"
                            }
                        },
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "$ref": "#/components/schemas/PreviewRequestRequest"
                            }
                        },
                        "multipart/form-data": {
                            "schema": {
                                "$ref": "#/components/schemas/PreviewRequestRequest"
                            }
                        }
                    },
                    "required": true
                },
                "security": [
                    {
                        "cookieAuth": []
                    }
                ],
                "responses": {
                    "201": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PreviewResponse"
                                }
                            }
                        },
                        "description": ""
                    },
                    "400": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorEnvelope"
                                }
                            }
                        },
                        "description": ""
                    }
                }
            }
        },
        "/api/v1/document-templates/{template_id}/publish/": {
            "post": {
                "operationId": "document_template_publish",
                "parameters": [
                    {
                        "in": "path",
                        "name": "template_id",
                        "schema": {
                            "type": "integer"
                        },
                        "required": true
                    }
                ],
                "tags": [
                    "document-templates"
                ],
                "security": [
                    {
                        "cookieAuth": []
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Template"
                                }
                            }
                        },
                        "description": ""
                    },
                    "409": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorEnvelope"
                                }
                            }
                        },
                        "description": ""
                    }
                }
            }
        },
        "/api/v1/document-templates/{template_id}/versions/": {
            "post": {
                "operationId": "document_template_version_create",
                "parameters": [
                    {
                        "in": "path",
                        "name": "template_id",
                        "schema": {
                            "type": "integer"
                        },
                        "required": true
                    }
                ],
                "tags": [
                    "document-templates"
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/TemplateVersionWriteRequest"
                            }
                        },
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "$ref": "#/components/schemas/TemplateVersionWriteRequest"
                            }
                        },
                        "multipart/form-data": {
                            "schema": {
                                "$ref": "#/components/schemas/TemplateVersionWriteRequest"
                            }
                        }
                    }
                },
                "security": [
                    {
                        "cookieAuth": []
                    }
                ],
                "responses": {
                    "201": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TemplateVersion"
                                }
                            }
                        },
                        "description": ""
                    },
                    "400": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorEnvelope"
                                }
                            }
                        },
                        "description": ""
                    }
                }
            }
        },
        "/api/v1/document-templates/{template_id}/versions/{version_id}/": {
            "patch": {
                "operationId": "document_template_version_update",
                "parameters": [
                    {
                        "in": "path",
                        "name": "template_id",
                        "schema": {
                            "type": "integer"
                        },
                        "required": true
                    },
                    {
                        "in": "path",
                        "name": "version_id",
                        "schema": {
                            "type": "integer"
                        },
                        "required": true
                    }
                ],
                "tags": [
                    "document-templates"
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PatchedTemplateVersionWriteRequest"
                            }
                        },
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "$ref": "#/components/schemas/PatchedTemplateVersionWriteRequest"
                            }
                        },
                        "multipart/form-data": {
                            "schema": {
                                "$ref": "#/components/schemas/PatchedTemplateVersionWriteRequest"
                            }
                        }
                    }
                },
                "security": [
                    {
                        "cookieAuth": []
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TemplateVersion"
                                }
                            }
                        },
                        "description": ""
                    },
                    "400": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorEnvelope"
                                }
                            }
                        },
                        "description": ""
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "AssetUploadRequest": {
                "type": "object",
                "properties": {
                    "file": {
                        "type": "string",
                        "format": "binary"
                    }
                },
                "required": [
                    "file"
                ]
            },
...
            "DocumentAsset": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "readOnly": true
                    },
                    "key": {
                        "type": "string",
                        "readOnly": true
                    },
                    "content_type": {
                        "type": "string",
                        "readOnly": true
                    },
                    "size": {
                        "type": "integer",
                        "readOnly": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "readOnly": true
                    }
                },
                "required": [
                    "content_type",
                    "created_at",
                    "id",
                    "key",
                    "size"
                ]
            },
            "DocumentLanguage": {
                "enum": [
                    "en",
                    "de"
                ],
                "type": "string",
                "description": "* `en` - en\n* `de` - de"
            },
...
            "ErrorEnvelope": {
                "type": "object",
                "properties": {
                    "error": {
                        "$ref": "#/components/schemas/ErrorBody"
                    }
                },
                "required": [
                    "error"
                ]
            },
...
            "FormatEnum": {
                "enum": [
                    "html",
                    "pdf"
                ],
                "type": "string",
                "description": "* `html` - html\n* `pdf` - pdf"
            },
...
            "PreviewRequestRequest": {
                "type": "object",
                "properties": {
                    "language": {
                        "$ref": "#/components/schemas/DocumentLanguage"
                    },
                    "invoice_id": {
                        "type": "integer",
                        "minimum": 1
                    },
                    "format": {
                        "allOf": [
                            {
                                "$ref": "#/components/schemas/FormatEnum"
                            }
                        ],
                        "default": "pdf"
                    }
                },
                "required": [
                    "language"
                ]
            },
            "PreviewResponse": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid",
                        "readOnly": true
                    },
                    "url": {
                        "type": "string",
                        "readOnly": true
                    },
                    "expires_at": {
                        "type": "string",
                        "format": "date-time",
                        "readOnly": true
                    }
                },
                "required": [
                    "expires_at",
                    "id",
                    "url"
                ]
            },
...
            "Template": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "readOnly": true
                    },
                    "code": {
                        "type": "string",
                        "readOnly": true,
                        "pattern": "^[-a-zA-Z0-9_]+$"
                    },
                    "name": {
                        "type": "string",
                        "readOnly": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "readOnly": true
                    },
                    "versions": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/TemplateVersion"
                        },
                        "readOnly": true
                    }
                },
                "required": [
                    "code",
                    "created_at",
                    "id",
                    "name",
                    "versions"
                ]
            },
...
            "TemplateVersionWriteRequest": {
                "type": "object",
                "properties": {
                    "language": {
                        "$ref": "#/components/schemas/DocumentLanguage"
                    },
                    "html": {
                        "type": "string"
                    },
                    "css": {
                        "type": "string"
                    },
                    "page_settings": {},
                    "asset_keys": {
                        "type": "array",
                        "items": {
                            "type": "string",
                            "minLength": 1
                        }
                    }
                }
            },
            "TemplateWriteRequest": {
                "type": "object",
                "properties": {
                    "code": {
                        "type": "string",
                        "minLength": 1,
                        "maxLength": 80,
                        "pattern": "^[-a-zA-Z0-9_]+$"
                    },
                    "name": {
                        "type": "string",
                        "minLength": 1,
                        "maxLength": 120
                    }
                },
                "required": [
                    "code",
                    "name"
                ]
            },
...
        },
        "securitySchemes": {
            "cookieAuth": {
                "type": "apiKey",
                "in": "cookie",
                "name": "sessionid"
            }
        }
    }
}